Feat: Add RecodeList style
