--- frontend/src/components/UI/organisms/GetQuery/style.ts
@@ -1,1 +1,0 @@
-export default {};
