+++ frontend/src/components/UI/organisms/RecodeList/index.stories.tsx
@@ -1,38 +1,56 @@
 import React from 'react';
 import GlobalThemeProvider from 'styles/GlobalThemeProvider';
-import RecodeList from '.';
+import RecodeList, { ITimeRecode } from '.';
 
 export default {
   title: 'Organisms/RecodeList',
   component: RecodeList,
 };
 
-const temp = [
+const TestData = [
   {
-    title: 'lifemanager',
-    startTime: { hour: 11, min: 30 },
-    endTime: { hour: 12, min: 30 },
+    title: 'lifemanager 개발',
+    startTime: {
+      hour: 10,
+      min: 20,
+    },
+    endTime: {
+      hour: 13,
+      min: 0,
+    },
     category: 'develop',
     isActive: true,
   },
   {
-    title: '낮잠',
-    startTime: { hour: 11, min: 30 },
-    endTime: { hour: 12, min: 30 },
-    category: 'sleep',
+    title: 'pkiop.me 개발',
+    startTime: {
+      hour: 14,
+      min: 0,
+    },
+    endTime: {
+      hour: 20,
+      min: 0,
+    },
+    category: 'develop',
+    isActive: true,
+  },
+  {
+    title: '쉬기',
+    startTime: {
+      hour: 21,
+      min: 0,
+    },
+    endTime: {
+      hour: 22,
+      min: 0,
+    },
+    category: 'rest',
     isActive: false,
   },
-  {
-    title: '부캠',
-    startTime: { hour: 11, min: 30 },
-    endTime: { hour: 12, min: 30 },
-    category: 'develop',
-    isActive: true,
-  },
-];
+] as ITimeRecode[];
 
 export const Default = () => (
   <GlobalThemeProvider>
-    <RecodeList recodeList={temp} />
+    <RecodeList timeRecodes={TestData} loading={false} error={null}/>
   </GlobalThemeProvider>
 );
