+++ frontend/src/components/UI/organisms/RecodeList/style.ts
@@ -1,8 +1,13 @@
 import styled from 'styled-components';
-import RecodeComponent from 'components/UI/molecules/Recode';
+import { Recode } from 'components/UI/molecules/Recode/style';
 
-export const RecodeList = styled.div``;
-
-export const Recode = styled(RecodeComponent)``;
+export const RecodeList = styled.div`
+  margin-top: 1rem;
+  & > ${Recode} {
+    :not(:last-child) {
+      margin-bottom: 0.4rem;
+    }
+  }
+`;
 
 export default {};
